spark/requirements.txt
pyspark
vaderSentiment
textblob
python-dotenv
nltk
pytz
pandas
pyarrow